requirements.txt
# Core AI & ML
langchain
langgraph
crewai

# Vector Database & Embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.2

# Document Processing
PyPDF2>=3.0.1
python-docx>=0.8.11
openpyxl>=3.1.2
markdown>=3.5.1

# LLM Local Integration
llama-cpp-python>=0.2.0

# Web Framework
streamlit>=1.28.0
gradio>=3.50.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0

# NLP Dependencies
nltk>=3.8
textblob>=0.17.1

# Development
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0

# Additional fixes
mammoth>=1.6.0
